+++ lib/easy_app.dart
@@ -4,7 +4,7 @@ import 'package:easy_app/screen/base_screen.dart';
 import 'package:easy_app/screen/main_screen.dart';
 import 'package:easy_app/screen/page_manager.dart';
 import 'package:easy_app/utils/languages.dart';
-import 'package:easy_app/utils/network_util.dart';
+import 'package:easy_app/utils/network_utils.dart';
 import 'package:easy_app/utils/os.dart';
 import 'package:flutter/material.dart';
 import 'package:path_provider/path_provider.dart';
@@ -16,8 +16,9 @@ class EasyApp {
     List<String> languages = const ["en_US"],
     required BaseScreen homeScreen,
   }) async {
-    if (OS.getOS() != OSType.web)
+    if (OS.getOS() != OSType.web) {
       localPath = (await getApplicationDocumentsDirectory()).path;
+    }
     if (activateConnectionChecker) NetworkUtils.init();
     if (languages.isNotEmpty) await Language.init(languages);
     PageManager.init(homeScreen);